requirements.txt
# Core ML/NLP
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2
faiss-cpu==1.7.4

# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# API/Web
Flask==2.3.2
Flask-CORS==4.0.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0

# Optional: For BERT-LoRA
# peft==0.4.0

# Optional: For Jupyter notebooks
# jupyter==1.0.0
# matplotlib==3.7.2
# seaborn==0.12.2

# Development
# pytest==7.4.0
# flake8==6.0.0